Initial simple neural net model for testing
Before spending too much time on producing the best model possible,
first prove that the overall workflow is functional.  This model
is just a single neuron (and input normalization) but demonstrates
both that the training script work and that the output model can be
used to drive the Udacity simulator

Driving quality is understandably poor.  More sophisticated models
to come