Implement data generator; Stays on full track 1
Data generator is implemented to avoid having to store entire dataset
in memory during training (not an issue yet but can be with more data
augmentation).  No data augmentation is currently performed, not even
horizontal flipping of images from earlier versions

Saved model performs well enough to keep the car on track 1 for a
full lap.  While this is sufficient for submitting the project as
part of the Udacity Self Driving Car Nanodegree, I would like to
develop and tune the model further.  Model does not work well on the
second track

TODO:
* Data augmentation
* Dropout and Regularization in the model
* Tune model architecture
* Tune hyperparameters
* Produce more training data