Add data augmentation and visualization
Data generation and augmentation are split into their own file,
augment.py, for clarity

Data augmentation techniques implemented:
* Horizontal flip
* Translation
* Gaussian Noise
* Shadows
* Brightness

All the augmentations are randomized.  No augmentation is applied to
the validation dataset

Graphic visualization of the dataset, comprised of random dataset
images and the distribution of labeled steering angles, are saved
to disk.  Example images of augmentation are also output